+++ sdk/azcore/arm/runtime/policy_bearer_token_test.go
@@ -172,6 +172,7 @@ func TestBearerPolicy_GetTokenFailsNoDeadlock(t *testing.T) {
 }
 
 func TestBearerTokenWithAuxiliaryTenants(t *testing.T) {
+	t.Skip("this feature isn't implemented yet")
 	srv, close := mock.NewTLSServer()
 	defer close()
 	srv.AppendResponse(mock.WithBody([]byte(accessTokenRespSuccess)))
@@ -186,8 +187,8 @@ func TestBearerTokenWithAuxiliaryTenants(t *testing.T) {
 	b := NewBearerTokenPolicy(
 		mockCredential{},
 		&armpolicy.BearerTokenOptions{
-			Scopes:           []string{scope},
-			AuxiliaryTenants: []string{"tenant1", "tenant2", "tenant3"},
+			Scopes: []string{scope},
+			//AuxiliaryTenants: []string{"tenant1", "tenant2", "tenant3"},
 		},
 	)
 	pipeline := newTestPipeline(&azpolicy.ClientOptions{Transport: srv, Retry: retryOpts, PerRetryPolicies: []azpolicy.Policy{b}})